+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed 
+- Hook up Students Guide API [#313](https://github.com/rokwire/illinois-app/issues/313).
 
 ## [2.4.25] - 2021-06-17
 ### Changed